Move AppBar into common
