+++ src/App.scss
@@ -1,4 +1,4 @@
-@import './components/AppBar/variables';
+@import './components/common/AppBar/variables';
 @import './App_variables.scss';
 
 $page-gutter-size: 48px;
