+++ src/App_variables.scss
@@ -1,4 +1,4 @@
-@import './components/AppBar/variables';
+@import './components/common/AppBar/variables';
 $app-main-real-vert-padding: 24px;
 
 $app-remaining-viewport-height-for-content: calc(
