+++ src/components/common/AppBar/variables.scss
